requirements.txt
Django==1.11
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
inflection==0.3.1
olefile==0.44
Pillow==4.1.1
py==1.4.34
pytest==3.1.2
pytz==2017.2
sqlparse==0.2.3
